cycling calculator

+++ src/main/java/com/iver/records/Processor.java
@@ -2,6 +2,6 @@ package com.iver.records;
 
 public record Processor(
         int coreNumber,
-        float speed
+        int speed
 ) {
 }
